Implement pooling of parameters across offsets
Create time-bins to pool parameters. To retrieve a parameter, pull from the nearest bin.
When updating parameters, compute parameters for all offsets inside each bin.
	modified:   src/Regions/Region.cpp
	modified:   src/Regions/Region.h
	modified:   src/Configurations/Configuration.cpp
	modified:   src/Configurations/Configuration.h
	modified:   src/Configurations/Default.cpp

Fix bugs where offsets were declared as ints, not floats.
	modified:   src/Key.h
	modified:   src/ParameterIos/Netcdf.cpp

Use a configuration's 'region' option if set
	modified:   src/Run.cpp

+++ src/Key.h
@@ -118,7 +118,7 @@ namespace Key {
                std::string iConfigurationName,
                int iIndex);
          Component::Type mType;
-         int mOffset;
+         float mOffset;
          int mDate;
          int mInit;
          int mLocationId;
